prepare 8.0.0 release (#177)
## [8.0.0] - 2023-10-16
The latest version of this SDK supports the ability to manage migrations
or modernizations, using migration flags. You might use this
functionality if you are optimizing queries, upgrading to new tech
stacks, migrating from one database to another, or other similar
technology changes. Migration flags are part of LaunchDarkly's Early
Access Program. This feature is available to all LaunchDarkly customers
but may undergo additional changes before it is finalized.

For detailed information about this version, refer to the list below.
For information on how to upgrade from the previous version, read the
[migration
guide](https://docs.launchdarkly.com/sdk/server-side/dotnet/migration-7-to-8).

### Added:
- A new `Migration` type which provides an out-of-the-box configurable
migration framework.
- For more advanced use cases, added new `MigrationVariation` and
`TrackMigration` methods on LdClient.
- `ApplicationInfo`, for configuration of application metadata that may
be used in LaunchDarkly analytics or other product features. This does
not affect feature flag evaluations.
- Added `IsOffline` method to `ILdClient`. This was previously only
available on the concrete `LdClient` implementation.

### Removed:
- Remove support for `User` in `LdClient` methods. The
`Context.FromUser` method can be used to convert a `User` to a
`Context`. In a future version it may be removed.

### Changed:
- Upgraded `LaunchDarkly.EventSource` to `5.1.0`. This will change the
minimum `LaunchDarkly.Logging` version to `2.0.0` for all dependencies.

---------

Co-authored-by: Eli Bishop <eli@launchdarkly.com>
Co-authored-by: Ben Woskow <48036130+bwoskow-ld@users.noreply.github.com>
Co-authored-by: LaunchDarklyCI <LaunchDarklyCI@users.noreply.github.com>
Co-authored-by: LaunchDarklyCI <dev@launchdarkly.com>
Co-authored-by: Ember Stevens <ember.stevens@launchdarkly.com>
Co-authored-by: ember-stevens <79482775+ember-stevens@users.noreply.github.com>
Co-authored-by: LaunchDarklyReleaseBot <launchdarklyreleasebot@launchdarkly.com>
Co-authored-by: Ryan Lamb <4955475+kinyoklion@users.noreply.github.com>
Co-authored-by: Louis Chan <lchan@launchdarkly.com>
Co-authored-by: ld-repository-standards[bot] <113625520+ld-repository-standards[bot]@users.noreply.github.com>
Co-authored-by: Kane Parkinson <93555788+kparkinson-ld@users.noreply.github.com>